requirements.txt
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
argcomplete==3.6.2
boto3==1.38.3
botocore==1.38.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cohere==5.15.0
colorama==0.4.6
databricks-sdk==0.52.0
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.115.12
fastavro==1.10.0
fastmcp==2.5.1
filelock==3.18.0
fsspec==2025.3.2
google-auth==2.39.0
griffe==1.7.3
groq==0.23.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.56
langchain-openai==0.3.14
langsmith==0.3.37
logfire-api==3.14.1
markdown-it-py==3.0.0
mcp==1.9.1
mdurl==0.1.2
mistralai==1.7.0
openai==1.76.0
openapi-pydantic==0.5.1
opentelemetry-api==1.32.1
orjson==3.10.16
packaging==24.2
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-ai==0.1.6
pydantic-ai-slim==0.1.6
pydantic-evals==0.1.6
pydantic-graph==0.1.6
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.16.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
